# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VMSmith API
  version: 0.2.0
  description: |
    OpenAPI description for the VMSmith REST API.

    This spec covers the implemented `/api/v1/*` surface exposed by the daemon,
    including VM lifecycle, bulk actions, cloning, snapshots, images, templates,
    host discovery, quotas, and log retrieval.
servers:
  - url: http://localhost:8080/api/v1
    description: Local daemon
security:
  - bearerAuth: []
...
paths:
  /vms:
    get:
      tags: [VMs]
      summary: List virtual machines
      parameters:
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: VM list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VM'
        default:
          $ref: '#/components/responses/APIError'
    post:
      tags: [VMs]
      summary: Create a VM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMSpec'
      responses:
        '201':
          description: VM created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        '429':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /vms/stats/top:
    get:
      tags: [VMs]
      summary: Top VMs by metric
      description: |
        Returns the top-N VMs ranked by the requested metric, sorted descending.
        Values are taken from each VM's most recent in-memory metric sample;
        VMs without a sample for the requested metric are skipped. Returns 503
        when metrics collection is disabled.
      parameters:
        - name: metric
          in: query
          schema:
            type: string
            enum: [cpu, mem, disk_read, disk_write, net_rx, net_tx]
            default: cpu
          description: Which metric to rank by.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 5
        - name: state
          in: query
          schema:
            type: string
            enum: [running, all]
            default: running
          description: Limit to running VMs (default) or include all VMs.
      responses:
        '200':
          description: Top VMs ranked by metric
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopVMResponse'
        '400':
          $ref: '#/components/responses/APIError'
        '503':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /vms/bulk:
    post:
      tags: [VMs]
      summary: Run a bulk VM action
      description: Start, stop, or delete multiple VMs in one request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkVMActionRequest'
      responses:
        '200':
          description: Per-VM bulk action results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkVMActionResponse'
        '400':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /vms/{vmID}:
    get:
      tags: [VMs]
      summary: Get a VM
      parameters:
        - $ref: '#/components/parameters/VMID'
      responses:
        '200':
          description: VM detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '404':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
    patch:
      tags: [VMs]
      summary: Update a VM
      parameters:
        - $ref: '#/components/parameters/VMID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMUpdateSpec'
      responses:
        '200':
          description: Updated VM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          $ref: '#/components/responses/APIError'
        '404':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        '429':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
    delete:
      tags: [VMs]
      summary: Delete a VM
      parameters:
        - $ref: '#/components/parameters/VMID'
      responses:
        '204':
          description: VM deleted
        '404':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
...
  /images:
    get:
      tags: [Images]
      summary: List images
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Image list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
        default:
          $ref: '#/components/responses/APIError'
    post:
      tags: [Images]
      summary: Export a VM to an image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
      responses:
        '201':
          description: Image created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/APIError'
        '404':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /images/upload:
    post:
      tags: [Images]
      summary: Upload a qcow2 image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                  description: Optional display name override.
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        '507':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /images/{imageID}:
    delete:
      tags: [Images]
      summary: Delete an image
      parameters:
        - $ref: '#/components/parameters/ImageID'
      responses:
        '204':
          description: Image deleted
        '404':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /images/{imageID}/download:
    get:
      tags: [Images]
      summary: Download an image file
      parameters:
        - $ref: '#/components/parameters/ImageID'
      responses:
        '200':
          description: Binary image download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /templates:
    get:
      tags: [Templates]
      summary: List reusable VM templates
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Template list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VMTemplate'
        default:
          $ref: '#/components/responses/APIError'
    post:
      tags: [Templates]
      summary: Create a reusable VM template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMTemplate'
        '400':
          $ref: '#/components/responses/APIError'
        '413':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /templates/{templateID}:
    delete:
      tags: [Templates]
      summary: Delete a template
      parameters:
        - $ref: '#/components/parameters/TemplateID'
      responses:
        '204':
          description: Template deleted
        '404':
          $ref: '#/components/responses/APIError'
        default:
          $ref: '#/components/responses/APIError'
  /host/interfaces:
    get:
      tags: [Host]
      summary: Discover host interfaces
      responses:
        '200':
          description: Host interfaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HostInterface'
        default:
          $ref: '#/components/responses/APIError'
  /host/stats:
    get:
      tags: [Host]
      summary: Get host utilization stats
      responses:
        '200':
          description: Host stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostStats'
        default:
          $ref: '#/components/responses/APIError'
  /quotas/usage:
    get:
      tags: [Quotas]
      summary: Get current quota usage
      responses:
        '200':
          description: Quota usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaUsage'
        default:
          $ref: '#/components/responses/APIError'
  /logs:
    get:
      tags: [Logs]
      summary: List recent daemon logs
      parameters:
        - name: level
          in: query
          schema:
            type: string
          description: Minimum log level.
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Return entries at or after this timestamp.
        - name: source
          in: query
          schema:
            type: string
          description: Filter by structured log source.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Alias retained for compatibility.
      responses:
        '200':
          description: Log entries
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        default:
          $ref: '#/components/responses/APIError'
  /events:
    get:
      tags: [Events]
      summary: List recent events
      description: |
        Returns the most recent events matching the provided filters, sorted
        newest-first. Each event carries a monotonic numeric ID assigned by the
        EventBus, plus a structured payload (source, severity, attributes).

        Pagination is page/per_page; clients can also provide `until=<seq>` to
        fetch events older than a known sequence ID (cursor pagination).
      parameters:
        - name: vm_id
          in: query
          schema:
            type: string
          description: Filter by VM ID (exact match).
        - name: type
          in: query
          schema:
            type: string
          description: Filter by event type (e.g., `vm.started`).
        - name: source
          in: query
          schema:
            type: string
            enum: [libvirt, app, system]
          description: Filter by event source.
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warn, error]
          description: Filter by severity.
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Return only events with `occurred_at` after this RFC3339 timestamp.
        - name: until
          in: query
          schema:
            type: string
          description: Cursor — exclude events with sequence ID >= this uint64.
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Event list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        default:
          $ref: '#/components/responses/APIError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  headers:
    XTotalCount:
      description: Total matching records before pagination.
      schema:
        type: integer
  parameters:
    VMID:
      name: vmID
      in: path
      required: true
      schema:
        type: string
    ImageID:
      name: imageID
      in: path
      required: true
      schema:
        type: string
    TemplateID:
      name: templateID
      in: path
      required: true
      schema:
        type: string
...
    TagFilter:
      name: tag
      in: query
      schema:
        type: string
      description: Case-insensitive tag filter.
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
        enum: [creating, running, stopped, deleted, unknown]
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
  responses:
    APIError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    VMState:
      type: string
      enum: [creating, running, stopped, deleted, unknown]
    NetworkMode:
      type: string
      enum: [nat, macvtap, bridge]
    NetworkAttachment:
      type: object
      properties:
        name:
          type: string
        mode:
          $ref: '#/components/schemas/NetworkMode'
        host_interface:
          type: string
        bridge:
          type: string
        static_ip:
          type: string
        gateway:
          type: string
        mac_address:
          type: string
    VMSpec:
      type: object
      required: [name, image]
      properties:
        name:
          type: string
        image:
          type: string
        template_id:
          type: string
        cpus:
          type: integer
        ram_mb:
          type: integer
        disk_gb:
          type: integer
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        cloud_init_file:
          type: string
        ssh_pub_key:
          type: string
        default_user:
          type: string
        networks:
          type: array
          items:
            $ref: '#/components/schemas/NetworkAttachment'
        nat_static_ip:
          type: string
        nat_gateway:
          type: string
    VMUpdateSpec:
      type: object
      properties:
        cpus:
          type: integer
        ram_mb:
          type: integer
        disk_gb:
          type: integer
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        nat_static_ip:
          type: string
        nat_gateway:
          type: string
    VM:
      type: object
      required: [id, name, spec, state, disk_path, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          $ref: '#/components/schemas/VMSpec'
        state:
          $ref: '#/components/schemas/VMState'
        ip:
          type: string
        nat_mac:
          type: string
        disk_path:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BulkVMActionRequest:
      type: object
      required: [action, ids]
      properties:
        action:
          type: string
          enum: [start, stop, delete]
        ids:
          type: array
          minItems: 1
          items:
            type: string
    BulkVMActionResult:
      type: object
      required: [id, success]
      properties:
        id:
          type: string
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
...
    BulkVMActionResponse:
      type: object
      required: [action, results]
      properties:
        action:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkVMActionResult'
...
    TopVMItem:
      type: object
      required: [vm_id, name, state, value]
      properties:
        vm_id:
          type: string
        name:
          type: string
        state:
          type: string
        value:
          type: number
          format: double
        sample:
          $ref: '#/components/schemas/MetricSample'
    TopVMResponse:
      type: object
      required: [metric, limit, state, items]
      properties:
        metric:
          type: string
          enum: [cpu, mem, disk_read, disk_write, net_rx, net_tx]
        limit:
          type: integer
        state:
          type: string
          enum: [running, all]
        items:
          type: array
          items:
            $ref: '#/components/schemas/TopVMItem'
...
    Image:
      type: object
      required: [id, name, path, size_bytes, format, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        size_bytes:
          type: integer
          format: int64
        format:
          type: string
        source_vm:
          type: string
        created_at:
          type: string
          format: date-time
    CreateImageRequest:
      type: object
      required: [vm_id, name]
      properties:
        vm_id:
          type: string
        name:
          type: string
    VMTemplate:
      type: object
      required: [id, name, image, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
        cpus:
          type: integer
        ram_mb:
          type: integer
        disk_gb:
          type: integer
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        default_user:
          type: string
        networks:
          type: array
          items:
            $ref: '#/components/schemas/NetworkAttachment'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTemplateRequest:
      type: object
      required: [name, image]
      properties:
        name:
          type: string
        image:
          type: string
        cpus:
          type: integer
        ram_mb:
          type: integer
        disk_gb:
          type: integer
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        default_user:
          type: string
        networks:
          type: array
          items:
            $ref: '#/components/schemas/NetworkAttachment'
    HostInterface:
      type: object
      required: [name, ips, mac, is_up, is_physical]
      properties:
        name:
          type: string
        ips:
          type: array
          items:
            type: string
        mac:
          type: string
        is_up:
          type: boolean
        is_physical:
          type: boolean
    HostResourceUsageSummary:
      type: object
      required: [used, total]
      properties:
        used:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        available:
          type: integer
          format: int64
        percentage:
          type: integer
    HostStats:
      type: object
      required: [vm_count, cpu, ram, disk, active_sse_streams]
      properties:
        vm_count:
          type: integer
        cpu:
          $ref: '#/components/schemas/HostResourceUsageSummary'
        ram:
          $ref: '#/components/schemas/HostResourceUsageSummary'
        disk:
          $ref: '#/components/schemas/HostResourceUsageSummary'
        active_sse_streams:
          type: integer
          format: int64
          description: Number of currently-open Server-Sent Events streams (events + future metrics streams).
    QuotaUsageSummary:
      type: object
      required: [used]
      properties:
        used:
          type: integer
        limit:
          type: integer
    QuotaUsage:
      type: object
      required: [vms, cpus, ram_mb, disk_gb]
      properties:
        vms:
          $ref: '#/components/schemas/QuotaUsageSummary'
        cpus:
          $ref: '#/components/schemas/QuotaUsageSummary'
        ram_mb:
          $ref: '#/components/schemas/QuotaUsageSummary'
        disk_gb:
          $ref: '#/components/schemas/QuotaUsageSummary'
    LogField:
      type: object
      additionalProperties:
        type: string
    LogsResponse:
      type: object
      required: [entries, total]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        total:
          type: integer
    LogEntry:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        level:
          type: string
        source:
          type: string
        msg:
          type: string
        fields:
          $ref: '#/components/schemas/LogField'
...
    Event:
      type: object
      required: [id, type, source, severity, message, occurred_at]
      properties:
        id:
          type: string
          description: Stringified monotonic uint64 assigned by the EventBus.
        type:
          type: string
          description: Event type (e.g., `vm.started`, `vm.created`).
        source:
          type: string
          enum: [libvirt, app, system]
        vm_id:
          type: string
        resource_id:
          type: string
        severity:
          type: string
          enum: [info, warn, error]
        message:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
        actor:
          type: string
        occurred_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
          description: Legacy timestamp; prefer `occurred_at` for new code.